refactor(ui): give pure logic its own modules, not component files
react-refresh/only-export-components fires when a module mixes components
with anything else, because Vite can only hot-replace a module whose whole
export surface is components. In 26 places the rule was reading a real
mistake: pure logic — route targets, reconcilers, parsers, class recipes —
was parked in a component module, so it could only be tested by importing
the component that happened to host it.

Each move follows the convention its own directory already had:

  - settings/models/ already keeps pure siblings as lowercase `.ts`
    (eligibility, format, reorder, supply), so vendorMeta drops its unused
    `.tsx`, and modelHubRoutes / migrationScan / runtimeLifecycle join them.
  - context/ already splits I/O-free contract from provider (showPageDrag,
    unsavedChangesContext), so dockDoc and apiErrorParse follow.
  - ui/ already keeps a non-component module beside its components
    (field.ts), so badge-variants / button-variants / vault-request-session
    follow.
  - workbench/ already has harnessRuns / harnessAgents, so harnessTabs and
    sessionAgentName join them, and agentDisplayName moves next to the
    catalog loader it answers for.

No behavior changes: every moved definition is byte-identical, and the
call sites now import it from one owner instead of reaching through a
component. Three tests that imported logic through a page now import the
module directly; `MigrationBanner.test.ts` was already only testing the
scan gate, so it is renamed to `migrationScan.test.ts` after the subject.

vendorMeta is the clearest case of the rule reading a mistake in the other
direction: the file holds no JSX at all, but `export const DEFAULT_VENDOR =
VENDOR_OPTIONS[0]` is an uppercase export the rule cannot see through, so
it treated it as a component and flagged all 8 of its siblings.

Errors: 281 -> 255. react-refresh: 42 -> 16.
