+++ ui/src/components/apps/AppWindow.tsx
@@ -10,7 +10,8 @@ import { ShowPageAvatarContent } from '../../apps/showPageAvatarTile';
 import { useWindowManager, type WindowInstance } from '../../context/WindowManagerContext';
 import { useUnsavedChangesActionGuard } from '../../context/useUnsavedChangesActionGuard';
 import { clampToLayer, resizeBounds, type ResizeDir } from '../../lib/windowBounds';
-import { WindowBodyGestureShield, shouldShieldWindowBody } from './windowGesture';
+import { WindowBodyGestureShield } from './WindowBodyGestureShield';
+import { shouldShieldWindowBody } from './windowGesture';
 import { ErrorBoundary } from '../ui/error-boundary';
 import { ShowPageAnnotateControl } from '../workbench/ShowPageAnnotateControl';
 import { useShowPageAnnotationHost } from '../workbench/ShowPageAnnotationHostContext';
