+++ ui/src/components/settings/models/vendorMeta.ts
