+++ ui/src/components/ui/vault-request-session-link.tsx
@@ -1,31 +1,7 @@
 import { Link } from 'react-router-dom';
 
-import type { VaultRequest } from '@/context/ApiContext';
 import { cn } from '@/lib/utils';
-
-export type VaultRequestSessionDisplay = {
-  id: string;
-  label: string;
-  isWorkbench: boolean;
-  isIdFallback: boolean;
-};
-
-export function vaultRequestSessionDisplay(request: VaultRequest): VaultRequestSessionDisplay | null {
-  const card = (request.card ?? {}) as { session_id?: unknown };
-  const cardSessionId = typeof card.session_id === 'string' && card.session_id.trim() ? card.session_id.trim() : null;
-  const session = request.session ?? null;
-  const id = session?.id?.trim() || cardSessionId;
-  if (!id) return null;
-  const title = session?.title?.trim();
-  const sessionLabel = session?.label?.trim();
-  const label = title || sessionLabel || id;
-  return {
-    id,
-    label,
-    isWorkbench: Boolean(session?.is_workbench),
-    isIdFallback: label === id && !title && (!sessionLabel || sessionLabel === id),
-  };
-}
+import type { VaultRequestSessionDisplay } from './vault-request-session';
 
 export const VaultRequestSessionLink: React.FC<{
   session: VaultRequestSessionDisplay;
