+++ ui/src/components/workbench/ShowPageShareControl.tsx
@@ -8,7 +8,8 @@ import { Popover, PopoverContent, PopoverTrigger } from '../ui/popover';
 import { Select } from '../ui/select';
 import { Switch } from '../ui/switch';
 import { useApi } from '../../context/ApiContext';
-import { showDockId, useDock } from '../../context/DockContext';
+import { useDock } from '../../context/DockContext';
+import { showDockId } from '../../context/dockDoc';
 import { copyTextToClipboard } from '../../lib/utils';
 import { isIosDevice, isRealMobileSafari, isStandalonePwa } from '../../lib/platform';
 import { copyHref, type ShowPageLinkInfo } from '../../lib/showPageLinks';
