+++ ui/src/components/workbench/VaultsPage.tsx
@@ -17,7 +17,8 @@ import { useToast } from '../../context/ToastContext';
 import type { ApprovalOutcome } from '../ui/vault-approval-card';
 import { SigningAddressList } from '../ui/signing-address-list';
 import { VaultApprovalDialog } from '../ui/vault-approval-dialog';
-import { VaultRequestSessionLink, vaultRequestSessionDisplay } from '../ui/vault-request-session-link';
+import { vaultRequestSessionDisplay } from '../ui/vault-request-session';
+import { VaultRequestSessionLink } from '../ui/vault-request-session-link';
 import { VaultSecretDialog } from '../ui/vault-secret-dialog';
 import { VaultSettingsDialog } from '../ui/vault-settings-dialog';
 import { useProtectedVault } from '../../lib/useProtectedVault';
